DLModels/StarTrek/requirements.txt
# --- Core Deep Learning ---
tensorflow==2.17.0
numpy==1.26.0
keras==3.13.2

# --- Image Processing ---
Pillow==12.1.0

# --- Deployment / UI ---
streamlit==1.54.0

# --- Utilities ---
pandas==2.3.3
matplotlib==3.9.2
seaborn==0.13.2